feat: add links to home page and change router
