+++ src/components/Sidebar.js
@@ -1,4 +1,5 @@
 import { useSelector } from "react-redux";
+import { Link } from "react-router-dom";
 
 const Sidebar = () => {
 	const openMenu = useSelector((store) => store.sidebar.openMenu);
@@ -9,15 +10,17 @@ const Sidebar = () => {
 				<div className="grid grid-cols-1 bg-gray-200">
 					<div className="flex flex-col gap-5 py-5 px-8 mt-4">
 						<ul className="flex flex-col gap-5">
-							<li className="font-bold cursor-pointer">Home</li>
 							<li className="font-bold cursor-pointer">
-								Subscriptions
+								<Link to="/">Home</Link>
 							</li>
 							<li className="font-bold cursor-pointer">
-								History
+								<Link to="/">Live</Link>
 							</li>
 							<li className="font-bold cursor-pointer">
-								Library
+								<Link to="/">Subscriptions</Link>
+							</li>
+							<li className="font-bold cursor-pointer">
+								<Link to="/">History</Link>
 							</li>
 						</ul>
 					</div>
