+++ src/index.js
@@ -1,14 +1,40 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import React from "react";
+import ReactDOM from "react-dom/client";
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+
+import "./index.css";
+
+import App from "./App";
+import Body from "./components/Body";
+import VideoPage from "./components/watch/VideoPage";
+
+import reportWebVitals from "./reportWebVitals";
+
+const routerConfig = [
+	{
+		path: "/",
+		element: <App />,
+		children: [
+			{
+				index: true,
+				element: <Body />,
+			},
+			{
+				path: "/watch",
+				element: <VideoPage />,
+			},
+		],
+	},
+];
+
+const router = createBrowserRouter(routerConfig);
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+	<React.StrictMode>
+		<RouterProvider router={router} />
+	</React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
